Switch to next diff in the context of the main thread
This makes the UI much more responsive than doing it in the context of the
IO thread and the overall reactivity shouldn't be negatively impacted even if
no next diff is yet available thanks to coroutines. Even if it was, it wouldn't
be a huge deal as there is little reason to interact with the UI without a diff.